Extended fix for xlim/ylim initialization.
xlim and ylim are initialized within an if statement.  No default
values provided.  Initializing to some dummy value and testing
for that value after the initialization attempt would be
sufficient if I knew what the possible valid values for xlim and
ylim could be.  Instead, added an extra initialization variable
for each and tested those instead.