Rewrite of tf_nc_cmr_tiled_is_valid()
Untested code.

The original threw compiler warnings based on potential buffer
overflows with sprintf().  On further inspection, the code
allowed multiple nonconfirming files to be reported as valid.

I have refactored the code to use malloc() to handle the buffer
overflow issues and removed most of the nesting.  These are
intrusive changes compared to the rest of the work I've been
doing and really should have another set of eyes.